refactor: remove a classe Address que não é mais necessária.
